Refactor pyfuncserver project structure
