+++ python/pyfunc-server/benchmark/benchmark.sh
@@ -16,19 +16,19 @@ sleep 5
 echo "=================================="
 echo "============small payload========="
 echo "=================================="
-cat benchmark/small.cfg | vegeta attack -rate 100 -duration=60s | vegeta report
+cat small.cfg | vegeta attack -rate 100 -duration=60s | vegeta report
 
 echo "=================================="
 echo "============medium payload========="
 echo "=================================="
-cat benchmark/medium.cfg | vegeta attack -rate 100 -duration=60s | vegeta report
+cat medium.cfg | vegeta attack -rate 100 -duration=60s | vegeta report
 
 echo "=================================="
 echo "============large payload========="
 echo "=================================="
-cat benchmark/large.cfg | vegeta attack -rate 100 -duration=60s | vegeta report
+cat large.cfg | vegeta attack -rate 100 -duration=60s | vegeta report
 
 echo "=================================="
 echo "===========max throughput========="
 echo "=================================="
-cat benchmark/large.cfg | vegeta attack -rate 0 -max-workers=8 -duration=60s | vegeta report
+cat large.cfg | vegeta attack -rate 0 -max-workers=8 -duration=60s | vegeta report
