+++ python/pyfunc-server/pyfuncserver/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .model import PyFuncModel
+from pyfuncserver.model.model import PyFuncModel
