+++ python/pyfunc-server/pyfuncserver/__main__.py
@@ -23,6 +23,7 @@ from prometheus_client import CollectorRegistry, multiprocess
 
 from pyfuncserver import PyFuncModel
 from pyfuncserver.server import PyFuncServer
+from pyfuncserver.utils.contants import ERR_DRY_RUN
 
 DEFAULT_MODEL_NAME = "model"
 DEFAULT_LOCAL_MODEL_DIR = "/tmp/model"
@@ -50,37 +51,13 @@ if __name__ == "__main__":
 
     # load model
     model = PyFuncModel(args.model_name, args.model_dir)
+
     try:
         model.load()
     except Exception as e:
         logging.error(f"Unable to initalize model")
         logging.error(traceback.format_exc())
-        logging.error(
-            """
-            Some module might be missing from your pyfunc model.
-            If the module is part of your code, 
-            you can specify it as 'code' argument when calling log_pyfunc, e.g.:
-            
-            merlin.log_pyfunc_model(model_instance=MyModel(), 
-                    conda_env="env.yaml",
-                    code=["path/to/your/package"])
-                    
-            If the module is a python package, then you should add it in the
-            conda env.yaml that you pass to log_pyfunc method, e.g.:
-            
-            dependencies:
-            - python=3.7
-            - pip:
-              - my-missing-package==1.0
-              
-            OR if the package is not available in PyPI distribution
-            
-            dependencies:
-            - python=3.7
-            - my-missing-package=1.0
-            
-            See: https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-from-an-environment-yml-file
-            """)
+        logging.error(ERR_DRY_RUN)
         exit(1)
 
     if args.dry_run:
